Make bpf_get_mem global
Also revise `bpf_get_memem()` to differentiate between invalid memory range and write-protected region
