+++ bpf/include/bpf.h
@@ -148,6 +148,7 @@ int bpf_install_hook(bpf_t *bpf);
 void bpf_add_region(bpf_t *bpf, bpf_mem_region_t *region,
                     void *start, size_t len, uint8_t flags);
 
+void* bpf_get_mem(const bpf_t *bpf, uint8_t size, const intptr_t addr, uint8_t type);
 int bpf_store_allowed(const bpf_t *bpf, void *addr, size_t size);
 int bpf_load_allowed(const bpf_t *bpf, void *addr, size_t size);
